Rename files and package to be more descriptive
