+++ tensorflow_graphics/projects/mesh_rcnn/ops/cubify.py
